requirements.txt
pygame==2.5.2
nltk==3.8.1
textblob==0.17.1
Pillow==10.3.0
vaderSentiment==3.3.2
requests==2.31.0
